+++ config/settings/cb_status.yaml
@@ -49,3 +49,10 @@ STAGE:
 # クラバトカテゴリーの場所
 CATEGORY:
   POSITION: 9
+
+# bot操作のメッセージID
+BOT_OPERATION:
+  CONVEX_SITUATION: '885842181569536020'
+  TIME_LIMIT: '885850652448788491'
+  ATTENDANCE: '885865380625678387'
+  TASK_KILL: '885856546578374718'
