--- config/settings/cb_time.yaml
@@ -1,33 +1,0 @@
-# 活動限界時間
-TIME_LIMIT_EMOJI:
-  EMOJI:
-    _0: '818660609607467110'
-    _1: '818660731351597066'
-    _2: '818660755947126826'
-    _3: '818660775731658793'
-    _4: '818660795813986374'
-    _5: '818660810556571650'
-    _6: '818660819826114590'
-    _7: '818660830685298718'
-    _8: '818661029923127296'
-    _9: '818661040363536394'
-    _10: '818661052647866379'
-    _11: '818661061321818112'
-    _12: '818661075494240316'
-    _13: '818661083262484531'
-    _14: '818661091575463986'
-    _15: '818661099980587029'
-    _16: '818661110592831489'
-    _17: '818661118876713021'
-    _18: '818661130326114325'
-    _19: '818661140171063358'
-    _20: '818661150429937705'
-    _21: '818661159942750249'
-    _22: '818661168833888316'
-    _23: '818661178179059806'
-  DISPLAY: '821323690565828622'
-  FIRST: '821323691962793994'
-  LATTER: '821323693212041216'
-
-# 離席中のメッセージID
-ATTENDANCE: '835782794037821470'
