+++ config/settings/channel.yaml
@@ -24,7 +24,7 @@ CHANNEL_ID:
   # 進行-連携
   PROGRESS: '720773327395881130'
   # 活動時間
-  ACTIVITY_TIME: '795629249754431488'
+  BOT_OPERATION: '881962172673060894'
   # 持越凸先
   CARRYOVER_DESTINATION: '761073612261818378'
