+++ src/client/interactionCreate.ts
@@ -4,6 +4,7 @@ import ThrowEnv from 'throw-env'
 import * as declare from '../convex/declare/interaction'
 import * as limit from '../convex/time/limit'
 import * as manage from '../convex/manage'
+import * as kill from '../convex/role/kill'
 
 /**
  * インタラクションが起きた際の処理を実行する
@@ -22,5 +23,8 @@ export const InteractionCreate = async (interaction: Discord.Interaction) => {
   if (comment) return console.log(comment)
 
   comment = await limit.Interaction(interaction)
+  if (comment) return console.log(comment)
+
+  comment = await kill.Interaction(interaction)
   if (comment) return console.log(comment)
 }
