+++ src/client/ready.ts
@@ -1,7 +1,7 @@
 // import * as Discord from 'discord.js'
 import Settings from 'const-settings'
 import {Client} from '../index'
-import * as react from '../convex/react'
+import * as etc from '../convex/etc'
 import * as util from '../util'
 
 /**
@@ -12,26 +12,7 @@ export const Ready = async () => {
   await channel.send('きゃるきゃるーん')
 
   // リアクションのキャッシュを取る
-  react.Fetch()
-
-  /*
-  {
-    const channel = util.GetTextChannel('881962172673060894')
-
-    channel.send('離席中状態の変更')
-
-    const text = [
-      '<@&797189384595832853> は、このメッセージがオレンジ色になります。',
-      '↓のボタンで離席中状態を変更できます。',
-    ].join('\n')
-    const 離席中 = new Discord.MessageButton().setCustomId('riseki-on').setStyle('PRIMARY').setLabel('離席中')
-    const 解除 = new Discord.MessageButton().setCustomId('riseki-off').setStyle('SECONDARY').setLabel('解除')
-    await channel.send({
-      content: text,
-      components: [new Discord.MessageActionRow().addComponents(離席中).addComponents(解除)],
-    })
-  }
-  // */
+  etc.Fetch()
 
   console.log(`Logged in as ${Client.user?.username}!`)
 }
