+++ src/convex/format.ts
@@ -2,8 +2,8 @@ import * as Discord from 'discord.js'
 import moji from 'moji'
 import Option from 'type-of-option'
 import Settings from 'const-settings'
-import * as util from '../../util'
-import {TLList, TLFormat} from '../../util/type'
+import * as util from '../util'
+import {TLList, TLFormat} from '../util/type'
 
 /**
  * TLを正しい書式に整形させる、
