+++ tsconfig.json
@@ -31,19 +31,17 @@
     "src/command/clanbattle.ts",
     "src/command/index.ts",
     "src/command/management.ts",
+    "src/convex/category.ts",
+    "src/convex/etc.ts",
+    "src/convex/format.ts",
     "src/convex/lapAndBoss.ts",
     "src/convex/manage.ts",
     "src/convex/over.ts",
-    "src/convex/react.ts",
-    "src/convex/role.ts",
     "src/convex/situation.ts",
     "src/convex/declare/index.ts",
     "src/convex/declare/list.ts",
     "src/convex/declare/react.ts",
     "src/convex/declare/status.ts",
-    "src/convex/etc/category.ts",
-    "src/convex/etc/format.ts",
-    "src/convex/etc/index.ts",
     "src/convex/plan/delete.ts",
     "src/convex/plan/edit.ts",
     "src/convex/plan/index.ts",
@@ -52,6 +50,8 @@
     "src/convex/report/cancel.ts",
     "src/convex/report/index.ts",
     "src/convex/report/update.ts",
+    "src/convex/role/index.ts",
+    "src/convex/role/kill.ts",
     "src/convex/time/attendance.ts",
     "src/convex/time/limit.ts",
     "src/io/bossTable.ts",
